perform integration with sparkworker scala sources
