--- java/logging.scala
@@ -1,29 +1,0 @@
-package SparkWorker
-
-import java.text.SimpleDateFormat
-import java.util.Calendar
-import java.util.Date
-
-object Logging {
-  def getDate() : String = {
-    val now = Calendar.getInstance().getTime()
-    val logFormat = new SimpleDateFormat("yy/MM/dd HH:mm:ss")
-    return logFormat.format(now)
-  }
-
-  def log(message: String) = {
-    System.out.println(getDate() + " INFO sparklyr-worker: " + message)
-  }
-
-  def logError(message: String) = {
-    System.err.println(getDate() + " ERROR sparklyr-worker: " + message)
-  }
-
-  def logError(message: String, e: Exception) = {
-    System.err.println(getDate() + " ERROR sparklyr-worker: " + message, e.toString)
-  }
-
-  def logWarning(message: String) = {
-    System.err.println(getDate() + " WARN sparklyr-worker: " + message)
-  }
-}
