+++ java/spark-1.5.2/fileutils.scala
@@ -1,4 +1,4 @@
-package SparkWorker
+package sparklyr
 
 import java.io.File
 import java.util.UUID
