+++ java/spark-1.5.2/rdd.scala
@@ -1,4 +1,4 @@
-package SparkWorker
+package sparklyr
 
 import org.apache.spark._
 import org.apache.spark.rdd.RDD
@@ -9,59 +9,6 @@ import scala.reflect.ClassTag
 import sparklyr.Backend
 import sparklyr.Logger
 import sparklyr.JVMObjectTracker
-
-class WorkerContext[T: ClassTag](
-  rdd: RDD[T],
-  split: Partition,
-  task: TaskContext,
-  lock: AnyRef,
-  closure: Array[Byte],
-  columns: Array[String]) {
-
-  private var result: Array[T] = Array[T]()
-
-  def getClosure(): Array[Byte] = {
-    closure
-  }
-
-  def getColumns(): Array[String] = {
-    columns
-  }
-
-  def getSourceIterator(): Iterator[T] = {
-    rdd.iterator(split, task)
-  }
-
-  def getSourceArray(): Array[T] = {
-    getSourceIterator.toArray
-  }
-
-  def getSourceArrayLength(): Int = {
-    getSourceIterator.toArray.length
-  }
-
-  def getSourceArraySeq(): Array[Seq[Any]] = {
-    getSourceArray.map(x => x.asInstanceOf[Row].toSeq)
-  }
-
-  def setResultArray(resultParam: Array[T]) = {
-    result = resultParam
-  }
-
-  def setResultArraySeq(resultParam: Array[Any]) = {
-    result = resultParam.map(x => Row.fromSeq(x.asInstanceOf[Array[_]].toSeq).asInstanceOf[T])
-  }
-
-  def getResultArray(): Array[T] = {
-    result
-  }
-
-  def finish(): Unit = {
-    lock.synchronized {
-      lock.notify
-    }
-  }
-}
 
 class WorkerRDD[T: ClassTag](
   parent: RDD[T],
@@ -157,18 +104,5 @@ class WorkerRDD[T: ClassTag](
     logger.log("Wait using lock for RScript completed")
 
     return workerContext.getResultArray().iterator
-  }
-}
-
-object WorkerHelper {
-  def computeRdd(df: DataFrame, closure: Array[Byte]): RDD[Row] = {
-
-    val parent: RDD[Row] = df.rdd
-    val computed: RDD[Row] = new WorkerRDD[Row](
-      parent,
-      closure,
-      df.columns)
-
-    computed
   }
 }
